news_briefing_api/requirements.txt
crawl4ai
flask
schedule
requests
beautifulsoup4
sumy
nltk
yake
python-dotenv


